Prevent test in production (#447)
Closes https://github.com/datagouv/data.gouv.fr/issues/1727 

This PR now uses the `warning` form from our validation system with a
new `testNotAllowed` validation rule.
It moves the warning from the accordion to under the field, like the
banner from the issue.
It removes the boldness of the banners to lower their visual importance.
It moves the organization form from vuelidate to our own `useForm`.
It removes the `i18n` utils from cdata because it's not necessary
anymore.