Prevent sending too many resources in parallel (#502)
This PR is linked to https://github.com/opendatateam/udata/pull/3338
(but do not require it)

Without this PR, the udata PR will work but will print "failing to save
resource after x retry" because we send too many files in a short time
and the retry strategy doesn't manage to find a window to update the
resource. By chunking we reduce this problem.