feat: JSON preview until 1MB JSON files (#586)
After discussing with @geoffreyaldebert , we decided to increase the
size limit of JSON files that can have a JSON preview to 1,000,000
characters instead of previously 100,000 chars (~1MB instead of ~100KB).

JSON preview is partly collapsed by default, so that should not degrade
user experience - which is not the case with XML preview.