fix: load preview if remote whatever the CORS status, and display error if CORS issue (#589)
## Problem

Issue and possible solutions are described in detail here:
https://github.com/datagouv/data.gouv.fr/issues/1810

For example, it's the case of all PDFs on the page:
-
http://data.gouv.fr/datasets/service-public-fr-annuaire-de-l-administration-base-de-donnees-locales/

## Enable remote file previews, with CORS error handling as an error msg

Load preview component even if the file is hosted on remote server and
display error if CORS issue.
This would allow for more files to have a preview **BUT** it means that
in case of a CORS issue, user would have a disappointing "Aperçu" tab
with an error message, instead of not displaying the "Aperçu" tab at all
as the previous behaviour.
Do we want that?

This look like that:
<img width="1234" height="215" alt="screenshot 2025-07-17 at 18 14 02"
src="https://github.com/user-attachments/assets/8cc932c3-d7f6-45ec-8e66-d89af9c84deb"
/>

As I understand it, the `hasPreview` computed property cannot be async
since Vue computed properties must be synchronous, while CORS checking
requires async operations (fetch requests), so we cannot check for CORS
issue in the parent component to avoid displaying the "Aperçu" tab at
all.

## Fix URL inconsistencies

In this PR we also fix some inconsistencies, always using stable URL
`resource.latest` instead of `resource.url`.