Fix contact points in form (#597)
Requires https://github.com/opendatateam/udata/pull/3376

It fixes the translation issue (we can't use unescaped `@` anymore in
translation because it has a special meaning for vue-i18n).

Also it adds a new system of "subform" to allow validation of contact
points when the parent form is validating.

This is a bit cumbersome because we must only validate new contact
points and not existing ones.