Components release and related fixes (#636)
Most changes are imports moved to `@datagouv/components-next` or fixes
related to the components when used outside of `cdata`. The fixes are :
- missing packages from `@datagouv/components-next`
- missing components from `@datagouv/components-next`

Some types required in udata-front-kit were also moved to
`@datagouv/components-next` to avoid duplicates.

`staticUrl` is removed from both `cdata` and
`@datagouv/components-next`. It was only used for the oembed script but
now it is available from cdata itself.

`@datagouv/components-next` is type-checked before build. This is a best
practice and it is required in our case because we provide TS file (and
not type definitions + JS file) and `udata-front-kit` check them.

---------

Co-authored-by: Thibaud Dauce <ThibaudDauce@users.noreply.github.com>
Co-authored-by: Thibaud Dauce <git@hadibut.fr>