Fix too many requests in ObjectsSelect (#666)
Since there is a `object.id in objectsByIds.value`, `watchEffect` was
triggering again on `objectsByIds.value[object.id] =` but we don't need
to fetch on `objectsByIds` changes, only on `selectedSuggest` ones.

You can test on
http://dev.local:3000/admin/reuses/6721fde6ff0522e9cc7ef5d0/datasets

Maybe we could find a way to fetch all objects at once (by passing them
from the parent? Because the ObjectsSelect don't have a clear way to do
that, but the parent could do a `/api/1/datasest?reuse=xxx`, maybe?)