Fix toggletip button (#667)
Our current Toogletip is putting one button (our BrandedButton) inside
another button (the PopoverButton).

This is wrong for semantic, accessibility and lead to strange hover
issues.

I didn't found a way to use both components, because the PopoverButton
is expecting to be a button and not a Vue component (even if it has a
`as` prop to change it).