Basic new translation system (#696)
A simple `npm update` is breaking our usage of `vue-i18n` /
`@nuxtjs/i18n` (from 9.3 to 9.5).

We don't know why so I made a little experiment to build our own
translate function.

The keys files:
- **160 lines** `useTranslation.ts` expose the composable to have the
`t()` translate function
- **10 lines** `plugins/translation.ts` load the translations keys for
the current locale in a global state and expose the `$t()` for the
templates
- **50 lines** `TranslationT.vue` component alternative to put HTML in
translation slots
- **100 lines** `design/translations.vue` a small page to show the usage
of the function and the component
- **110 lines** `translation.spec.ts` some tests based on the design
page above
- **30 lines** `locale/en.json` a small example JSON file containing the
keys from the design page to test that the english version is working

The code is only 220 lines of code (and 240 lines of tests), so I think
it's a simple enough change to avoid configuring complex plugins… (for a
translation system we don't even use…)

It also removes 45 lines of obscure configuration.

Small summary:
- [x] New `useTranslation` replacing `useI18n()`
- [x] New `<TranslationT>` replacing `<i18n-t>`
- [x] Remove `DevOnly` on `design.vue` to allow E2E tests (with `npm run
build && npm run preview`)
- [x] Remove `meta` and `link` in `<head>` from I18n (not sure it's
needed)
- [x] Remove some code that called `localePath()` (to build with
prefix?) or `localeRoute()` to get the full route object.
- [x] `locale` is no longer a `Ref()` it's a simple `string`
- [x] `useTranslation` is much more stable than `useI18n`, it should be
possible to use it anywhere (even outside setups)
- [x] Fixed a weird problem following
https://github.com/datagouv/cdata/pull/691 where `navigateTo()` crashed
because it was called outside Nuxt context (maybe because it was after a
first await? But why crash in this PR? Maybe the removal of the
`node_modules/unplugin-vue-router` which was a dependency of `vue-i18n`)
- [ ] Need to change `NUXT_PUBLIC_I18N_BASE_URL` into
`NUXT_PUBLIC_BASE_URL`
- [x] Update doc 

---

Not sure if we want to keep these meta tags?
```
<meta property="og:url" content="https://www.data.gouv.fr/">
<meta property="og:locale" content="fr">
<meta property="og:locale:alternate" content="en">
<meta property="og:locale:alternate" content="es">
```

We already have `<html  lang="fr">`, it should be enough?